+++ tests/testthat/test-output.R
@@ -22,3 +22,24 @@ test_that("different outputs produce identical code", {
   expect_equal(out1, out2)
   expect_equal(out1, out2)
 })
+
+test_that("intermediate outputs are always valid svg", {
+  path <- tempfile()
+  svglite(path)
+
+  expect_valid_svg <- function() {
+    expect_error(xml2::read_xml(path), NA)
+  }
+
+  mini_plot(1:10)
+  expect_valid_svg()
+
+  rect(2, 2, 3, 3)
+  expect_valid_svg()
+
+  segments(5, 5, 6, 6)
+  expect_valid_svg()
+
+  dev.off()
+  expect_valid_svg()
+})